refactor: restructure event and topic query architecture
  - Consolidate EventListQueryResponse into EventFeedQueryResponse
  - Move services to appropriate domain layers for better organization
  - Simplify API by returning full event data instead of just IDs
  - Remove redundant EventQueryApi and merge functionality into feed endpoint
