+++ src/main/kotlin/com/flownews/api/event/api/EventFeedQueryApi.kt
@@ -18,14 +18,14 @@ class EventFeedQueryApi(
 ) {
     @Operation(
         summary = "사용자별 이벤트 피드 조회",
-        description = "인증된 사용자의 개인화된 이벤트 피드의 이벤트 ID 목록을 반환합니다. 각 이벤트의 상세 정보는 /events/{id} 엔드포인트로 별도 요청해야 합니다.",
+        description = "인증된 사용자의 개인화된 이벤트 피드를 상세 정보와 함께 반환합니다.",
     )
     @GetMapping("/api/events/feed")
     fun getUserEventFeed(
         @CurrentUser user: User?,
         @RequestParam(required = false) category: String?,
-    ): ResponseEntity<EventFeedQueryResponse> {
-        val eventIds = eventFeedQueryService.getEventFeeds(user, category).map { it.requireId() }
-        return ResponseEntity.ok(EventFeedQueryResponse(eventIds))
+    ): ResponseEntity<List<EventFeedQueryResponse>> {
+        val eventFeed = eventFeedQueryService.getEventFeeds(user, category)
+        return ResponseEntity.ok(eventFeed)
     }
 }
